Restructures embed -> core
